move SourceMap to its own file, add source_map arg in client for requesting source map to be generated
