+++ algosdk/__init__.py
@@ -16,5 +16,6 @@ from . import util
 from . import v2client
 from . import wallet
 from . import wordlist
+from . import source_map
 
 name = "algosdk"
